+++ backend/src/main/java/com/example/smu_club/club/controller/MemberClubController.java
@@ -1,55 +1,36 @@
-/*
 package com.example.smu_club.club.controller;
 
-import java.util.List;
+
 import com.example.smu_club.answer.dto.AnswerRequestDto;
-import com.example.smu_club.club.dto.*;
-import com.example.smu_club.common.ApiResponseDto;
+import com.example.smu_club.club.dto.ApplicationFormResponseDto;
+import com.example.smu_club.club.dto.ApplicationRequestDto;
+import com.example.smu_club.club.dto.ApplicationResponseDto;
 import com.example.smu_club.club.service.ClubService;
+import com.example.smu_club.common.ApiResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
+
+/**
+ * @author sjy
+ */
 @RestController
 @RequiredArgsConstructor
-public class ClubController {
+@RequestMapping("/api/v1/member")
+public class MemberClubController {
+
     private final ClubService clubService;
 
-    */
-/**
-     *
-     * GUEST 권한
-     *//*
-
-    @GetMapping("/api/v1/public/clubs")
-    public ResponseEntity<ApiResponseDto<List<ClubsResponseDto>>> findAllClubs() {
-        List<ClubsResponseDto> clubs = clubService.findAllClubs();
-
-        return ResponseEntity.ok(ApiResponseDto.success(clubs, "전체 클럽 목록 조회 성공 [메인페이지]"));
-    }
-
-    @GetMapping("/api/v1/public/clubs/{clubId}")
-    public ResponseEntity<ApiResponseDto<ClubResponseDto>> findClubById(@PathVariable Long clubId){
-        ClubResponseDto club = clubService.findClubById(clubId);
-
-        return ResponseEntity.ok(ApiResponseDto.success(club, "클럽 상세 정보 조회 성공"));
-    }
-
-
-
-    */
-/**
-     *
-     * MEMBER 권한
-     *//*
-
-    @GetMapping("/api/v1/member/clubs/{clubId}/apply")
-    public ResponseEntity<ApiResponseDto<ApplicationFormResponseDto>> getApplication (
+    @GetMapping("/clubs/{clubId}/apply")
+    public ResponseEntity<ApiResponseDto<ApplicationFormResponseDto>> getApplication(
             @PathVariable Long clubId,
             @AuthenticationPrincipal UserDetails userDetails
-            ){
+    ) {
 
         ApplicationFormResponseDto responseDto = clubService.findMemberInfoWithQuestions(clubId, userDetails);
 
@@ -61,12 +42,13 @@ public class ClubController {
         return ResponseEntity.ok(apiResponseDto);
     }
 
-    @PostMapping("/api/v1/member/clubs/{clubId}/apply")
-    public ResponseEntity<ApiResponseDto<ApplicationResponseDto>> createApplication (
+
+    @PostMapping("/clubs/{clubId}/apply")
+    public ResponseEntity<ApiResponseDto<ApplicationResponseDto>> createApplication(
             @PathVariable Long clubId,
             @AuthenticationPrincipal UserDetails userDetails,
             @RequestBody ApplicationRequestDto applicationRequestDto
-            ){
+    ) {
         List<AnswerRequestDto> ard = applicationRequestDto.getQna();
         String fileUrl = applicationRequestDto.getFileUrl();
 
@@ -78,8 +60,4 @@ public class ClubController {
 
         return ResponseEntity.ok(apiResponseDto);
     }
-
-
-
 }
-*/
